+++ CMakeLists.txt
@@ -24,7 +24,7 @@ set(Headers
     include/types/Categories.h
     include/types/DataGenerator.h
     include/types/AlternativesPerformance.h
-    include/types/ProfileUpdater.h
+    include/learning/ProfileUpdater.h
     include/learning/LinearSolver.h
     include/learning/WeightUpdater.h
     include/app.h
@@ -42,7 +42,7 @@ set(Sources
     src/types/Categories.cpp
     src/types/DataGenerator.cpp
     src/types/AlternativesPerformance.cpp
-    src/types/ProfileUpdater.cpp
+    src/learning/ProfileUpdater.cpp
     src/learning/LinearSolver.cpp
     src/learning/WeightUpdater.cpp
     src/app.cpp
